Moved DataResult, TimelineTweet and TimelineUser types to composite types folder
