+++ src/types/response/composite/DataResult.ts
